Allow creating bitsets larger than 2^32 elements in CEGB
I ran into an issue where I get
```
LightGBM/src/treelearner/cost_effective_gradient_boosting.hpp:62:52: runtime error: signed integer overflow: 3331 * 749988 cannot be represented in type 'int'
```
because my num_features * num_data > 2^32

Switching to size_t (goes up to 2^64 on 64 bit platforms) to alleviate this.

+++ include/LightGBM/utils/common.h
@@ -807,8 +807,8 @@ inline static void ObtainMinMaxSum(const T1 *w, int nw, T1 *mi, T1 *ma, T2 *su) 
   }
 }
 
-inline static std::vector<uint32_t> EmptyBitset(int n) {
-  int size = n / 32;
+inline static std::vector<uint32_t> EmptyBitset(size_t n) {
+  size_t size = n / 32;
   if (n % 32 != 0) ++size;
   return std::vector<uint32_t>(size);
 }
